app/backend/pyproject.toml
[tool.poetry]
name = "schedule"
version = "0.1.0"
description = ""
authors = ["Andrey Sorokozherdev <asorokozherdyev@gmail.com>"]
license = "MIT"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
django-redis = "^5.4.0"
fake-useragent = "^1.5.1"
flower = "^2.0.1"
gunicorn = "^23.0.0"
lxml = "^5.3.0"
beautifulsoup4 = "^4.12.3"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
redis = "^5.0.8"
requests = "^2.32.3"
celery = "^5.4.0"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
pytest-django = "^4.9.0"
model-bakery = "^1.20.0"
pytest = "^8.3.4"
faker = "^33.1.0"
pytest-mock = "^3.14.0"
freezegun = "^1.5.1"
drf-spectacular = "^0.28.0"
django-sslserver = "^0.22"
django-extensions = "^3.2.3"
werkzeug = "^3.1.3"
pyopenssl = "^25.0.0"
django-cors-headers = "^4.7.0"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.4.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
